Add Radio shuffle mode; Fix shuffle bug
This fixes the bug that when playing a Group (e.g. a folder) by long pressing and then pressing "play", always the first song was selected. Now, a random song is played.

It also adds radio shuffle. This is normal shuffle but starts any song at a random point of time.