Retag: Add basic decisions frame work
This adds Kotlin and Jetpack Compose. It also  adds the first, very basic UI for interaction with music library retagging and restructuring. Currently, it can rename songs with a "The" prefix, e.g. "The Chords - Sh-Boom.mp3" becomes "Chords, The - Sh-Boom". But the renaming logic has not been implemented yet. It is very WIP.